Split the big file containing multiple classes into multiple files
