+++ app/src/main/java/com/example/checkpoint/dto/Clouds.kt
@@ -1,0 +1,5 @@
+package com.example.checkpoint.dto
+
+data class Clouds (
+    val all: Long
+)
