+++ app/src/main/java/com/example/checkpoint/dto/Coord.kt
@@ -1,0 +1,6 @@
+package com.example.checkpoint.dto
+
+data class Coord (
+    val lon: Long,
+    val lat: Long
+)
