+++ app/src/main/java/com/example/checkpoint/dto/Main.kt
@@ -1,0 +1,19 @@
+package com.example.checkpoint.dto
+
+import com.beust.klaxon.Json
+
+data class Main (
+    val temp: Double,
+
+    @Json(name = "feels_like")
+    val feelsLike: Double,
+
+    @Json(name = "temp_min")
+    val tempMin: Double,
+
+    @Json(name = "temp_max")
+    val tempMax: Double,
+
+    val pressure: Long,
+    val humidity: Long
+)
