+++ app/src/main/java/com/example/checkpoint/dto/Sys.kt
@@ -1,0 +1,9 @@
+package com.example.checkpoint.dto
+
+data class Sys (
+    val type: Long,
+    val id: Long,
+    val country: String,
+    val sunrise: Long,
+    val sunset: Long
+)
