+++ app/src/main/java/com/example/checkpoint/dto/Weather.kt
@@ -1,0 +1,8 @@
+package com.example.checkpoint.dto
+
+data class Weather (
+    val id: Long,
+    val main: String,
+    val description: String,
+    val icon: String
+)
