+++ app/src/main/java/com/example/checkpoint/dto/WeatherAPI.kt
@@ -3,9 +3,6 @@ package com.example.checkpoint.dto
 //
 //   val welcome = Welcome.fromJson(jsonString)
 
-import com.beust.klaxon.Json
-
-
 data class WeatherAPI (
     val coord: Coord,
     val weather: List<Weather>,
@@ -22,48 +19,14 @@ data class WeatherAPI (
     val cod: Long
 )
 
-data class Clouds (
-    val all: Long
-)
 
-data class Coord (
-    val lon: Long,
-    val lat: Long
-)
 
-data class Main (
-    val temp: Double,
 
-    @Json(name = "feels_like")
-    val feelsLike: Double,
 
-    @Json(name = "temp_min")
-    val tempMin: Double,
 
-    @Json(name = "temp_max")
-    val tempMax: Double,
 
-    val pressure: Long,
-    val humidity: Long
-)
 
-data class Sys (
-    val type: Long,
-    val id: Long,
-    val country: String,
-    val sunrise: Long,
-    val sunset: Long
-)
 
-data class Weather (
-    val id: Long,
-    val main: String,
-    val description: String,
-    val icon: String
-)
 
-data class Wind (
-    val speed: Double,
-    val deg: Long,
-    val gust: Double
-)
+
+
