+++ app/src/main/java/com/example/checkpoint/dto/Wind.kt
@@ -1,0 +1,7 @@
+package com.example.checkpoint.dto
+
+data class Wind (
+    val speed: Double,
+    val deg: Long,
+    val gust: Double
+)
